app: Add 'flip-back' action
This action is the counterpart of the 'flip' action. It's intended to be
triggered when flipping from the toolbox app back to the original app.
The action notifies the toolbox window in case the window needs to apply
any changes; for example, in the case of eos-knowledge-lib apps we need
to restart the app. (The default window does nothing in response to this
notification. Subclasses are intended to override that behaviour.)

This also adds an extra parameter to the trigger.js script so that you
can either flip or flip-back from the command line.

https://phabricator.endlessm.com/T23625