framework: Add framework app hacking toolbox
This adds a control panel for hacking eos-knowledge-lib-based apps.
Currently they need to be whitelisted, and we start with the only
whitelisted app being com.endlessm.dinosaurs.en.

The UI elements in the control panel should work at this point, but
don't do anything yet, even when flipping back to the app.

https://phabricator.endlessm.com/T23626