Apply framework controls to framework apps
Adds the code to take the current state of the model, and generate
app.json, overrides.css files, and GResource files; and relaunch the app
using the Restart API from eos-knowledge-lib.

We keep all the custom modules in a separate GResource file and pass it
directly when we relaunch the app.

https://phabricator.endlessm.com/T23626