framework: Add codeview
This adds codeview.js as a common widget for use within multiple
toolboxes. It has a signal "should-compile" that is emitted 250 ms after
any new code is typed. Applications can connect to this signal, analyze
the code, and use the setCompileResults() API to note any errors or
warnings in the UI, which will show up as gutter marks and highlight the
lines they are on. Clicking on the gutter mark will pop up a popover
with an error message.

It has API for setting the text programmatically without triggering any
compilation, and cancelling a pending compilation.

https://phabricator.endlessm.com/T23902