codeview: Add findAssignmentLocation method
This adds a findAssignmentLocation method which scans the toplevel
statements in the AST for assignment statements, and returns the
location of the right-hand side of the assignment. We will use this to
implement value validity and bounds checking for the hackable variables.

In order to do this, we add an "ast" property and cache its value, since
parsing is expensive. We invalidate the cache whenever the text in the
buffer changes.

https://phabricator.endlessm.com/T23902