codeview: Add "Fix" button
If you specify both a start and end location and a "fixme" property on
the object that you pass to setCompileResults(), then a "Fix" button
will become visible in the popover. Clicking it will delete the text
between the start and end locations, and replace it by the text
specified by the "fixme" property.

You can use this, for example, to reset a hackable variable to its
default if it's set to an invalid value.

https://phabricator.endlessm.com/T23902