window: Implement an unlock state
This is temporary! It is intended to be replaced by global quest state.
This will currently not persist after the hack toolbox app shuts down,
but it is intended to persist.

This models the unlock state as an array of booleans for each app, all
initially false. Each element in the array represents one level of
locking. Toolboxes can call unlock() to unlock the next level that is
still locked, or getUnlockState() to examine the current state.

https://phabricator.endlessm.com/T23929