framework: Implement reset for toolbox
This implements the binding to reset everything to default values when
the reset button is clicked.

It would be better to have one source of truth for all the default
values, but that would require implementing a function to transform JS
values to the kinds of string/number values that we put in the code
view. For now we'll keep it separate, and fix it later.

https://phabricator.endlessm.com/T23957