framework: Add widget for popup menu with choices
This takes the existing popup menu for the logos, and refactors it into
a separate component which is available for all toolboxes. You can
construct it with a map of string IDs to property values, as well as the
widget class to create to represent each choice (e.g. Gtk.Label or
Gtk.Image), the property to set on each widget, and extra properties to
give each widget.

We will use this component for more stuff in the next commit.

https://phabricator.endlessm.com/T23977