framework: Implement "Hyperlinks" switch
This adds a switch with which you can turn off navigation in the app.
Previously navigation would simply be turned off if you had click
sounds, and on in all other cases. This gives you more fine-grained
control to do what you want.

In the code view, it is allowed to write `yes`, `no`, `on`, or `off` as
well as the normal `true` or `false` for the hyperlinks variable. These
alternative boolean constants are unfortunately not yet highlighted.

https://phabricator.endlessm.com/T23977