framework: Only copy sound packs to app if needed
Copying the sound files through the FD pipe to the relaunched app takes
a noticeable while. We should avoid copying the sound files if we don't
really need them.

This splits the sound files into separate GResource bundles. First of
all, if no sounds are selected, then no sounds are passed into the app's
sandbox at all, so that should be fast. If a sound pack is selected,
then only that sound pack is copied into the sandbox, so restarting the
app may still be a bit slower, but less so than when we always copied
all the sounds.

Without copying sounds, the restart is 100–150 ms faster than before.

https://phabricator.endlessm.com/T23990