framework: Show spinner while relaunching
This shows a spinner in place of the minimize button while relaunching
the app. It's not possible to reliably get a signal when the app is
fully launched, shown, and flipped, so unfortunately we just fake it by
showing the spinner for 5 seconds. (The DBus method already returns
after about 250 ms, and the flip takes much longer than that.)

https://phabricator.endlessm.com/T23990