framework: Don't relaunch the app if nothing changed
If nothing has changed since the first launch of the toolbox or the last
relaunch of the app, then don't relaunch the app; just flip back over.

Note, this is the simplest possible implementation and doesn't care
about the case where you change one of the parameters then change it
back to its original value.

https://phabricator.endlessm.com/T23991