locksManager: Add abstraction layer for GameStateService
This is a abstraction layer between the toolbox
and the GameStateService, and provides helpers
for most common operations.

The reason why this is this a separated class/module
is because we might need this in other cases as well
e.g. if we happen to have a more complex lock.

https://phabricator.endlessm.com/T24156