framework: Keep track of default values for each app
This adds a Defaults class that keeps track of the default values for
each app. A model class is now generated on the fly for each framework
app with the correct default values for each of the properties, and the
old model class has become RaModelBase, the parent class for the
generated model classes.

The Defaults instance is passed into each UI panel's constructor so that
it knows what to set its UI elements to. This means we can remove the
bit where we sync UI property defaults back to the model before binding
the model's properties, as that would never have worked with multiple
apps and one UI.

https://phabricator.endlessm.com/T24204