framework: Make Hackdex Chapter 1 hackable
This adds com.endlessm.Hackdex_chapter_one to all the switch statements
where it is needed. We add the CSS and YAML overrides to gen.js, with
the requisite changes for enabling all the hacking parameters that we do
in the Dinosaurs app.

Since the hackdex isn't based on a template, we unfortunately have to
copy its app.yaml into a place where it's available to autobahn inside
this flatpak, so that we can "import" it when overriding it.