framework: Audio player should deal with >6 cards
Since adding the Grid arrangement, we have a possibility of having more
than 6 cards. Currently, each soundpack has 6 sounds, so we need to
reuse earlier ones if there are more than 6 cards. We currently set the
maximum number of cards playing sounds to 100, but that will be improved
anyway when this is ported to use the sound server.

https://phabricator.endlessm.com/T24204