framework: Move sounds to sound server
This deletes the sounds from the toolbox app, so that they can instead
live inside the sound server.

The AudioPlayer module is no longer necessary, and the NoiseArrangement
base class now talks to the sound server. For this we need to add the
permission to the flatpak to talk to the sound server.

An advantage of this scheme is that we don't have to send the sound
packs as a GResource file through an FD pipe to the respawned app, so it
should restart more quickly when sounds are selected.

https://phabricator.endlessm.com/T24260