framework: Implement Hackdex Chapter 1 "corruption" quest
The global game state keeps track of the state of this quest. In the
toolbox for com.endlessm.Hackdex_chapter_one, we read this key and do
certain defaults differently if the quest is active.

The solution to the quest is to set the "main color" to anything but
transparent, so when flipping back to the app, we check if the quest has
been solved and if so, set the global game state key appropriately.

https://phabricator.endlessm.com/T24258