codeview: Add selection drag sound
A bit of searching did not show a signal specifically for the selection
bounds changing. So this is implemented by connecting to the 'mark-set'
signal and ignoring any marks that are not the insert or selection_bound
marks.

We keep track of the previous selection length, and play the sound if
the new selection length is 1 more or less than the previous length.

https://phabricator.endlessm.com/T24355