framework, fizzics: Add live updating sounds
This does a bit of magic with notify signals to play a sound when
changes in the codeview update the GUI, and a different sound when
changes in the GUI update the codeview.

In Fizzics we rate-limit the GUI update sound to once per 100 ms,
because otherwise it's emitted a LOT when dragging a slider. In the
framework apps this isn't necessary, as all the sliders have been
removed in the new designs.

https://phabricator.endlessm.com/T24355