toolbox: Style reset and minimize buttons
This styles the reset button according to the spec in the provided SVG,
and uses a symbolic icon extracted from the SVG as the button icon.

Since the button is larger than before, it makes the headerbar taller,
which increases the height of the minimize button as well. The minimize
button is not supposed to have any border or background according to the
specs, so we remove that as well.

The spec is still missing hover states. Those can be added later.

https://phabricator.endlessm.com/T24381