framework: Add new List card layout option
This adds a new option (with a new icon), for a "list" card layout,
which uses Arrangement.List for the arrangement.

In the case of Arrangement.List, we also need to use the Card.Title in
it, in the Dinosaurs app and probably others in the future too, or else
it will tend to request too much space from the app. This requires a bit
more complication in the YAML-generating code.

https://phabricator.endlessm.com/T24039