framework: Generate some YAML programmatically
The YAML generating code is starting to get messy with many string
manipulations that try to generate human-readable YAML. That is not
actually necessary, as the YAML is quickly converted into JSON not meant
for human consumption.

As the logic for what to put in the YAML gets more complicated, it
becomes more readable to write that logic in code, instead of conditional
string interpolation.

We take advantage of the fact that JSON is legal YAML.

https://phabricator.endlessm.com/T24039