framework: Make com.endlessm.encyclopedia.en hackable
With all the previous fixes, we now have the tools we need to make the
Encyclopedia app hackable with the framework hacking toolbox.

Since we include Card.Default when the arrangement is not
Arrangement.List, we also have to include the font size changes for
Card.Default from the Thematic template overrides.

Nonetheless, there are a lot of inconsistencies created when hacking
this app. For one, most of the changes are not immediately visible on
the front page when you flip back. Also, this uses Card.Default at many
sizes that it was not originally envisioned to be used at, and therefore
has no responsive styling for. I've just skipped those cases for now,
and let them exist without styling.

https://phabricator.endlessm.com/T24039