framework: Better victory conditions for the corruption quest
Previously we would check if the changed color was equal to the
corrupted color rgba(255, 255, 255, 0). This means you could change the
color to any other color, say pure red, with alpha=0, and still win the
quest even though the text was still invisible.

The new victory conditions are that you have to change the color to
something with alpha more than 5%.

https://phabricator.endlessm.com/T24870