codeview: Record error metrics
In order to gain insights into what kind of errors players often make in
the codeview, and what things they try in order to solve them, we record
a metric when an error message is shown. The payload includes the app ID
of the app being hacked, a user function ID (which is currently blank to
signify the main code panel), the full code, and a list of error
messages shown with their location information.

Subsequent payloads have an ed script indicating the diff of the code
from the previous event. Ed scripts, though esoteric, are probably the
diff format that takes up the fewest bytes.