toolbox: Make slider page sizes equal to step sizes
It looks like the page size is used as the scroll wheel delta for all
GtkRange instances, except scroll bars. Since it is easy to accidentally
send a scroll event on the target hardware's touchpad, we want the
scroll wheel delta to be the step size, not the page size, on all the
GtkScale instances that we have in our toolboxes.

The social force sliders in Fizzics already had a rather large step size
of 5, so additionally we reduce it to 2 here.

https://phabricator.endlessm.com/T24601