toolbox: Add new action to init the toolbox
The toolbox is being lazy loaded when the user clicks the FtH button. If
the toolbox is a bit complex, this loading can take some time and the
feedback for the user is odd because the interface freezes after the
click.

To try to *improve* the flipping time, we can preload the hack toolbox
for the app before the user clicks the FtH.

This patch adds a new action to create the toolbox widget so we can call
this from the shell and we can do the process transparent for the user.

https://phabricator.endlessm.com/T25093