System app: Add trap sequence in wobbly lockscreen
When the wobbly windows lockscreen is clicked, a trap sequence is
presented to the player, if running a specific quest. This sequence is
driven by the quest using the game state server as communication
channel.

If the player has the key needed to unlock the panel, the 'has-key'
background will be displayed as usual, but the player won't be able to
open it when clicked. Instead, game state 'tried' is set to the lock
for a few seconds. A quest can monitor this change to act
accordingly. There is a quest (Investigation) that presents a message,
telling the player to go away. And there is a quest (Breaking In) that
runs the trap sequence.

https://phabricator.endlessm.com/T25505