codeview: Add API to get the body of a toplevel function
Since we send only the bodies of user functions across the wire, we need
a way to obtain them without the surrounding "function foo() {" and "}".
We can't assume that any particular spaces or newlines will exist, so we
use the AST location information to extract the function body into a
string.

https://phabricator.endlessm.com/T25489