LightSpeed: Execute user functions with mock scope objects
We do this in order to catch errors such as misspelled identifiers
inside the function. (These are not resolved until the function is
executed, so they wouldn't be caught by just parsing the code.)

To do this, we need to mock the scope objects that the game passes when
executing the user functions, but with deterministic contents. These
mock scope objects need to be kept in sync with the game.

https://phabricator.endlessm.com/T25489