toolbox: Add left-hand column
Adds a left-hand column as specified in the new designs. Requires
passing the application ID into the toolbox constructor, in order to get
the application icon and name to display in the upper left-hand corner.

Removes the 'title' property of the toolbox, previously displayed in the
center of the header bar, as this is now no longer displayed anywhere.

Includes some extra changes to make the HackUnlock toolbox stay the same
as it was before.

https://phabricator.endlessm.com/T25567