toolbox: Hide headerbar along with toolbox panel
In the new design, we'll click on a topic to open the panel. The
headerbar should slide out along with it, and slide in when closing the
panel.

We still call the widget "headerbar", but it's just a regular box. It's
no longer styled as the headerbar of a window, so we don't need to tap
into those rules from Adwaita.

(This commit leaves things in an intermediate state where you can close
the panel but not open it again.)

https://phabricator.endlessm.com/T25481