toolbox: Peek in system's desktop files and icons for flipped app icon
The "masthead" of the toolbox needs to be able to look at the app info
of the flipped app. Unfortunately in most cases when running sandboxed,
we won't be able to use Gio.DesktopAppInfo directly for this. In order
for Gio.DesktopAppInfo to be loaded validly, the program listed in the
Exec line has to be found in the path, which it will not be for other
flatpaks. So, we try to load the Gio.DesktopAppInfo in the normal way
but if that fails we parse the .desktop file ourselves using code copied
from glib/gio/gdesktopappinfo.c.

https://phabricator.endlessm.com/T25567