toolbox: Add topics to left-hand menu
This adds the capability to split the toolboxes into topics. Topics
have a title and an icon name associated with them, and they appear in
the left-hand menu, and can scroll if there are too many to fit on the
screen at once.

The toolbox opens with the first topic selected. Clicking the minimize
button deselects the topic, and closes the panel. Selecting a topic
opens the panel if it is not yet open. The title widget at the top of
the panel displays the title and icon name of the currently showing
topic.

Topics can be marked as "needing attention" but that currently doesn't
do anything.

For now, all existing toolboxes use only one topic.

https://phabricator.endlessm.com/T25481