wobblyLockscreen: Play videos for the trapped sequence
Placeholder videos are used for now.

Use the "lock changed" callback to set the sequence state, which is then
used to play a different video in the _updateBackground method.

Overwrite the "play done" callback to play the last video in loop, and
end that video when the lock in finally unlocked.

https://phabricator.endlessm.com/T25505