utils: Improve workaround by which we get Gio.DesktopAppInfo
We need Gio.DesktopAppInfo in order to determine the user-visible name
and desktop icon of the app we're hacking. Gio.DesktopAppInfo will
consider itself invalid if it can't find the program listed in the
.desktop file's "Exec" line, so we need to work around that.

Previously we did this by reading in the .desktop file ourselves with
code adapted from Gio.DesktopAppInfo in GLib. A simpler way is to use
Gio.DesktopAppInfo itself, but before constructing the DesktopAppInfo,
override the "Exec" line in the .desktop file to some well-known binary
such as /bin/true. This workaround was copied from eos-discovery-feed.

https://phabricator.endlessm.com/T25567