framework: Align lockscreens vertically
On large resoluitions, the lockscreens were being vertically aligned
with GTK_ALIGN_FILL, making the widgets they covered taller than the
actual lockscreen assets. This changes the vertical alignment so that
the lockscreens properly cover the locked widgets.

Also sets a minimum height on the codeview, to ensure that it remains
the same height as it was in the previous toolbox design, until we can
refactor that design to fit the new toolbox better.

https://phabricator.endlessm.com/T25805