System: Adjust vertical alignment and W:H ratio of lockscreen
This is because the still image has background-size: cover, whereas the
video is rendered as if it had background-size: contain (and can't be
rendered any other way, as far as I can tell.) So if they are rendered
at any other width-to-height ratio, they will not overlap seamlessly.

This hardcodes the width-to-height ratio. It's not possible to use a
GtkAspectFrame here, since I can't figure out how to get it to work with
height-for-width allocation, which is what we actually want here.

The still image is still rendered at a different scale than the video,
so the transition is not yet seamless.

https://phabricator.endlessm.com/T25893