System: Create models at toolbox top level
In the other toolboxes, we create the models at the top level and bind
them to the UI panels in a bindModel() function. In the System toolbox,
we previously had the UI panels own their own models.

Since we have to write an unbindModel() function to solve the problem of
the bindings persisting after the toolbox is destroyed, we first
refactor the System toolbox to work like the other toolboxes.

https://phabricator.endlessm.com/T25961