Lightspeed: Merge the two spawn topics into one (backup plan)
We now have two topics in the Lightspeed toolbox that have something to
do with spawning, and we want to merge them into one because having two
different ones is confusing.

This is the easier, backup-plan, version of what we actually want to do:
have the two functions in the same code view. For this backup plan, we
put both code views in the same topic.

https://phabricator.endlessm.com/T26244