framework: Save decryption when relaunching the app
If we decrypt the app using the "rotation" property and flip back, then
the app will use our generated YAML file, so it will be decrypted.
However, if we then close the app and restart it, it will be back to
using the default YAML file and will be encrypted again.

To avoid this problem, we save the "rotation" property to the game state
when we flip back. On the hackdex side, if the "rotation" property is
not specified (i.e. we are using the default YAML), and a value is saved
in the game state, then we will use that value.

https://phabricator.endlessm.com/T25507