framework: Make Leviathan hackable
Leviathan (com.endlessm.Hackdex_chapter_two) is now hackable. This is a
bit broken compared to the Roster, Dinosaurs, or Encyclopedia support,
because we only intend to expose the level 2 hacking panel in Leviathan.
The app was designed such that none of the settings in the level 1 panel
really make sense to apply, so it will probably break if you try (e.g.
by using the master key to open levels 1 and 3.)

This is missing a few things; the Scramble and Zalgo text filters don't
work on the HTML pages, and the "Enable hyperlinks" and sound effects
settings don't work on the lightbox buttons.

https://phabricator.endlessm.com/T25507