framework: Override entire lightbox trees
We need to add properties to the lightboxes in order to make sounds and
"Enable hyperlinks" work on the lightbox elements on the home page of
Leviathan. Since the lightboxes are at the top of an entire hierarchy
of other elements, some of which we already overrode, we now need to
override these entire hierarchies.

For this purpose, we write the JSON to be overridden into a separate
file, since it is so large.

https://phabricator.endlessm.com/T25507