toolboxes: Fix check for the case where no variables were set
Most of the existing toolboxes included a check similar to:

    if (Object.getOwnPropertyNames(scope).every(prop => prop === null))
        return;

This code didn't do anything, since objects can't have null as a
property name.

What I originally intended this to do, was to check for the case where
no variables were set at all, for example if the user deleted all the
code in the codeview. However, that approach (even if it hadn't been
buggy) wouldn't have worked anyway, since we also have other names in
the scope that aren't variables.

To fix this, we define an array of valid variables, and if all of them
are null after the code is executed, then we return early. We ignore
any other names in the scope that aren't variables that the user can
set.

In Fizzics, we just delete the check. It doesn't make any sense to
perform it there, because the 'species' array will still be defined.

https://phabricator.endlessm.com/T26373