toolboxes: Replace string constants with actual constants
Typing quotes around values like 'invulnerable' has proven to cause
problems for players, since we don't actually teach them what strings
are.

Add constants to each app's scope, giving each one the value of the
string that would previously have been required, so that we can keep
backwards compatibility in the cases where code is saved.

In Fizzics the constants work slightly differently. We give them a
numeric value since that's what is passed to the game in the end.

https://phabricator.endlessm.com/T26373