codeview: Add another method to set compile results
We need a different method to set compile results from an exception
originating inside user code. An example use case from Sidetrack is
where an instruction like "riley.jump()" throws an exception in the case
where the instructions queue is already full. The exception is thrown in
the jump() method of the Riley object, which is in the toolbox code, so
in any case we don't want to get our location information from the
topmost exception frame.

In the above example, the exception stack looks something like

    jump@/path/to/userFunction.js:46:19
    instructions@/path/to/userFunction.js line 249 > Function:11:5
    _compile@/path/to/userFunction.js:290:13

The "line 249 > Function" part means a function created at runtime using
"new Function()" on line 249 of the surrounding file. (Such a function
created programmatically doesn't have any file information, so it uses
the file information of the surrounding code and appends that string to
it.)

Line 11, column 5 of the user function is where we want to mark the
exception in the codeview, so we look for the topmost stack frame
containing the string " > Function:" and retrieve the line and column
from there.

https://phabricator.endlessm.com/T26346