fizzics: When error searching, do not compare index with value
The codeview before used to display:
```
background = "space"
```

But because kids was not used to "", that was changed in 5ffe087 to
```
background = space
```

In the new change, that `space` variable stored the underlaying
value which is an index, in this case the index 1.

On error search, we were comparing the types of the string 'space'
with the index 1, and similarly for other cases. This is why, the
codeview displayed a several amount of errors marked in red.

This commit, modifies _searchForErrorWithProps to do a proper type
comparison. It not only disappears all the bunch of errors
appearing on the Fizzics codeview when `item.key.fizzics.2` and
`lock.fizzics.2` keys were granted, but also because there are no
more errors, no error sound is emitted anymore when launching
Fizzics.

https://phabricator.endlessm.com/T29464#805655